Adding some extra files

+++ 0x00-hello_world/6-size.c
@@ -13,7 +13,7 @@ int main(void)
 
 	long longtype;
 
-	long double longtypedouble;
+	long int longtypedouble;
 
 	float floattype;
